CR: type hints, move code around
