+++ tests/integration/test_apigateway.py
@@ -26,7 +26,7 @@ from localstack.services.apigateway.helpers import (
     connect_api_gateway_to_sqs,
     get_resource_for_path,
     get_rest_api_paths,
-    path_based_url,
+    path_based_url, import_api_from_openapi_spec,
 )
 from localstack.services.awslambda.lambda_api import add_event_source
 from localstack.services.awslambda.lambda_utils import (
@@ -40,7 +40,7 @@ from localstack.utils.aws import aws_stack
 from localstack.utils.common import clone, get_free_tcp_port, json_safe, load_file
 from localstack.utils.common import safe_requests as requests
 from localstack.utils.common import select_attributes, short_uid, to_str
-
+from localstack.utils.testutil import load_test_resource
 from .awslambda.test_lambda import (
     TEST_LAMBDA_LIBS,
     TEST_LAMBDA_NODEJS,
@@ -1742,3 +1742,52 @@ class TestAPIGateway(unittest.TestCase):
         apigw_client.create_deployment(restApiId=api_id, stageName="staging")
 
         return api_id
+
+
+def test_import_swagger_api():
+    api_spec = load_test_resource("openapi.swagger.json")
+    api_spec_dict = json.loads(api_spec)
+
+    backend = APIGatewayBackend(region_name="eu-west-1")
+    api_model = backend.create_rest_api(name="", description="")
+
+    imported_api = import_api_from_openapi_spec(api_model, api_spec_dict, {})
+
+    # test_cfn_handle_serverless_api_resource fails if we support title
+    # assert imported_api.name == api_spec_dict.get("info").get("title")
+    assert imported_api.description == api_spec_dict.get("info").get("description")
+
+    paths = {v.path_part for k, v in imported_api.resources.items()}
+    assert paths == {"/", "pets", "{petId}"}
+
+    resource_methods = {v.path_part: v.resource_methods for k, v in imported_api.resources.items()}
+    methods = {kk[0] for k, v in resource_methods.items() for kk in v.items()}
+    assert methods == {"POST", "OPTIONS", "GET"}
+
+    assert resource_methods.get("/").get("GET").method_responses == {
+        "200": {
+            "statusCode": "200",
+            "responseModels": None,
+            "responseParameters": {"method.response.header.Content-Type": "'text/html'"},
+        }
+    }
+
+    assert resource_methods.get("pets").get("GET").method_responses == {
+        "200": {
+            "responseModels": {
+                "application/json": {
+                    "items": {
+                        "properties": {
+                            "id": {"type": "integer"},
+                            "price": {"type": "number"},
+                            "type": {"type": "string"},
+                        },
+                        "type": "object",
+                    },
+                    "type": "array",
+                }
+            },
+            "responseParameters": {"method.response.header.Access-Control-Allow-Origin": "'*'"},
+            "statusCode": "200",
+        }
+    }
