+++ tests/unit/utils/test_common.py
@@ -281,9 +281,3 @@ def test_download_with_timeout():
     # clean up
     proxy.stop()
     rm_rf(tmp_file)
-
-
-def load_test_resource(file_name: str, file_path: str = None) -> str:
-    if file_path:
-        return load_file(os.path.join(os.path.dirname(__file__), file_path, file_name))
-    return load_file(os.path.join(os.path.dirname(__file__), "../templates", file_name))
